+++ src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -1,8 +1,8 @@
-package kitchenpos.menu.application;
+package kitchenpos.menugroup.application;
 
-import kitchenpos.menu.dto.MenuGroupRequest;
-import kitchenpos.menu.dto.MenuGroupResponse;
-import kitchenpos.menu.infra.MenuGroupRepository;
+import kitchenpos.menugroup.dto.MenuGroupRequest;
+import kitchenpos.menugroup.dto.MenuGroupResponse;
+import kitchenpos.menugroup.infra.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
