+++ src/main/java/kitchenpos/menugroup/domain/MenuGroupName.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menugroup.domain;
 
 import org.springframework.util.StringUtils;
