+++ src/main/java/kitchenpos/menugroup/infra/MenuGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.menu.infra;
+package kitchenpos.menugroup.infra;
 
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
