+++ src/test/java/kitchenpos/menugroup/application/MenuServiceTest.java
@@ -1,8 +1,9 @@
-package kitchenpos.menu.application;
+package kitchenpos.menugroup.application;
 
+import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.NotFoundMenuGroupValidator;
+import kitchenpos.menugroup.domain.NotFoundMenuGroupValidator;
 import kitchenpos.product.domain.Product;
 import kitchenpos.menu.domain.MenuPriceValidator;
 import kitchenpos.menu.dto.MenuProductRequest;
